Refactor index.js sub functions into controllers grouped by concern
Replace all usage or pluralz with Z
