+++ src/controllers/verification_controller.js
@@ -1,0 +1,8 @@
+async function respond(req, res) {
+  const { body } = req;
+  res.status(200).send(body.challenge)
+}
+
+module.exports = {
+  respond
+};
