+++ src/meta.js
@@ -1,0 +1,5 @@
+const meta = {
+  basedir: __dirname
+};
+
+ module.exports = meta;
