+++ src/z.js
@@ -3,6 +3,7 @@ const pluralize = require("pluralize");
 
 const ACCEPTED_TAG_TYPE = ["NNS", "NNPS"];
 const EMOJI_REGEXP = /:\w+:/
+const PLURALZ_REGEXP = /\b(\w{2,})z([.?!]*)\b/;
 
 const lexer = new pos.Lexer();
 // Add support for Slack emoji in lexing
@@ -22,9 +23,18 @@ function isPlural(word, tag) {
   return ACCEPTED_TAG_TYPE.includes(tag) && pluralize.isPlural(word);
 }
 
+function ignore(text) {
+  // No easy, reliable parsing of backticks, so just ignore it.
+  return /`/.test(text);
+}
+
 class Z {
   constructor(sentence) {
+    this.ignore = ignore(sentence);
+    if (this.ignore) { return; }
+
     this.taggedWords = this.tagWords(sentence)
+    this.matchesPluralz = Boolean(sentence.match(PLURALZ_REGEXP));
   }
 
   // Get POS tagging and keep track of whitespace after word
@@ -41,12 +51,23 @@ class Z {
   }
 
   hasPlurals() {
+    if (this.ignore) { return false; }
+
     return this.taggedWords.some(({ word, tag }) => {
       return ACCEPTED_TAG_TYPE.includes(tag) && pluralize.isPlural(word)
     });
   }
 
+  hasPluralz() {
+    if (this.ignore) { return false; }
+
+    // TODO: use tagged words in a similar manner.
+    return this.matchesPluralz;
+  }
+
   replace() {
+    if (this.ignore) { return null; }
+
     const wordz = this.taggedWords.map(({ word, tag, whitespace }) => {
       if (isPlural(word, tag)) {
         word = pluralize.singular(word) + "z";
